+++ CHANGELOG.md
@@ -8,6 +8,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 For a full diff see [`2.7.0...main`][2.7.0...main].
 
+### Added
+
+- Added `allRules` parameter to allow disabling and enabling all rules  ([#913]), by [@localheinz]
+
 ## [`2.7.0`][2.7.0]
 
 For a full diff see [`2.6.1...2.7.0`][2.6.1...2.7.0].
@@ -612,6 +616,7 @@ For a full diff see [`362c7ea...0.1.0`][362c7ea...0.1.0].
 [#902]: https://github.com/ergebnis/phpstan-rules/pull/902
 [#911]: https://github.com/ergebnis/phpstan-rules/pull/911
 [#912]: https://github.com/ergebnis/phpstan-rules/pull/912
+[#913]: https://github.com/ergebnis/phpstan-rules/pull/913
 
 [@cosmastech]: https://github.com/cosmastech
 [@enumag]: https://github.com/enumag